+++ 0_RPackages/R-3/R-3.4.4/R-3.4.4_toBeInstalled/1_libraryInstall.sh
@@ -1,6 +1,10 @@
 #!/bin/bash 
-nohup Rscript /scratch/libraryInstall.R > /scratch/nohup.out
-Rscript /scratch/dockerFileGenerator.R
+if ! nohup Rscript /scratch/libraryInstall.R > /scratch/nohup.out; then
+    exit 1
+fi
+if ! Rscript /scratch/dockerFileGenerator.R; then
+    exit 1
+fi
 rm -r /scratch/packages
 rm /scratch/nohup.out
 rm /scratch/toDownload.txt
