+++ 0_RPackages/R-3/R-3.4.4/runMe.sh
@@ -52,8 +52,10 @@ mkdir -p $pathSharedfoldDock
 mv Dockerfile_1 Dockerfile
 sync
 dockerTempName=$(echo "$1" | tr '[:upper:]' '[:lower:]')
-docker build . -t $dockerTempName
-
+if ! docker build . -t $dockerTempName; then
+    echo "Docker container failed!! check log"
+    exit 1
+fi
 retry cp -r . $pathSharedfoldDock
 sync
 rm $pathSharedfoldDock/Dockerfile*
@@ -61,14 +63,21 @@ sync
 retry cp configurationFile.R $pathSharedfoldDock/R-3.4.4_toBeInstalled/libraryInstall.R
 sync
 echo "Step : R library install. This might take some time."
-docker run -tv $pathSharedfoldHost/R-3.4.4_toBeInstalled:/scratch $dockerTempName /scratch/1_libraryInstall.sh
+if ! docker run -tv $pathSharedfoldHost/R-3.4.4_toBeInstalled:/scratch $dockerTempName /scratch/1_libraryInstall.sh; then
+    echo "Docker container failed!! check log"
+    exit 1
+fi
 mv Dockerfile Dockerfile_1
 sync
 mv Dockerfile_2 Dockerfile
 sync
 mv Dockerfile Dockerfile_2
 sync
-retry cp Dockerfile_2 $pathSharedfoldDock/Dockerfile
+cp Dockerfile_2 $pathSharedfoldDock/Dockerfile
 sync
+if ! docker build $pathSharedfoldDock/ -t $dockerTempName; then
+    echo "Docker container failed!! check log"
+    exit 1
+fi
 echo 'DockerFile generation is done. Locate in DockerFolder and build your final docker.\n You can remove the temporary docker with docker rmi '$dockerTempName
