+++ 2_GUI/Jupyter/jupyter_lab/RtoBeInstalled/1_libraryInstall.sh
@@ -1,7 +1,12 @@
 #!/bin/bash 
-nohup Rscript /scratch/libraryInstall.R > /scratch/nohup.out
-Rscript /scratch/dockerFileGenerator.R
+if ! nohup Rscript /scratch/libraryInstall.R > /scratch/nohup.out; then
+    exit 1
+fi
+if ! Rscript /scratch/dockerFileGenerator.R; then
+    exit 1
+fi
 rm -r /scratch/packages
 rm /scratch/nohup.out
 rm /scratch/toDownload.txt
 rm /scratch/toInstall.txt
+
